restructure centralize 2 apps
