+++ package.json
@@ -11,7 +11,9 @@
 	},
 	"dependencies": {
 		"@faker-js/faker": "^9.5.0",
+		"cross-env": "^7.0.3",
 		"dayjs": "^1.11.13",
+		"dotenv": "^16.5.0",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-import": "^2.29.1",
