+++ playwright.config.ts
@@ -1,15 +1,16 @@
 import { defineConfig, devices } from '@playwright/test';
+import dotenv from 'dotenv';
+import path from 'path';
 
-/**
- * Read environment variables from file.
- * https://github.com/motdotla/dotenv
- */
-// import path from 'path';
-// dotenv.config({ path: path.resolve(__dirname, '.env') });
+const currentEnvType = process.env.ENV || 'public';
+dotenv.config({
+	path: path.resolve(__dirname, './src/env', `${currentEnvType}.env`),
+});
+
 /**
  * See https://playwright.dev/docs/test-configuration.
  */
-export default  ({
+export default {
 	testDir: './tests',
 	/* Run tests in files in parallel */
 	fullyParallel: false,
@@ -27,14 +28,14 @@ export default  ({
 	use: {
 		/* Base URL to use in actions like `await page.goto('/')`. */
 		// baseURL: 'http://127.0.0.1:3000',
-    extraHTTPHeaders: {
-      'Cache-Control': 'no-cache',
-      'Pragma': 'no-cache',
-    },
+		extraHTTPHeaders: {
+			'Cache-Control': 'no-cache',
+			'Pragma': 'no-cache',
+		},
 		/* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
 		bypassCSP: true,
 		storageState: undefined,
-		viewport: { width: 2000, height: 1000},
+		viewport: { width: 2000, height: 1000 },
 		trace: 'on',
 		headless: false, //for running on github action. run cmd line by addition --headed
 		permissions: ['camera', 'microphone'],
@@ -89,10 +90,10 @@ export default  ({
 		//   name: 'Microsoft Edge',
 		//   use: { ...devices['Desktop Edge'], channel: 'msedge' },
 		// },
-		 {
-		   name: 'Google Chrome',
-		   use: { ...devices['Desktop Chrome'], channel: 'chrome' },
-		 },
+		{
+			name: 'Google Chrome',
+			use: { ...devices['Desktop Chrome'], channel: 'chrome' },
+		},
 	],
 
 	/* Run your local dev server before starting the tests */
@@ -101,4 +102,4 @@ export default  ({
 	//   url: 'http://127.0.0.1:3000',
 	//   reuseExistingServer: !process.env.CI,
 	// },
-});
+};
