--- src/constant/env.ts
@@ -1,3 +1,0 @@
-//export const WEBSITE_URL = 'bldrdojo.com/apps/?app=asset';
-// FOR DEVEN export
- export const WEBSITE_URL = 'http://localhost/apps/?app=asset';
