+++ src/pages/assets_app/modules/common/components/detail_page.ts
@@ -1,5 +1,5 @@
 import { expect, Page } from '@playwright/test';
-import { BasePage } from '../../../base_page';
+import { BasePage } from '../../../../base_page';
 
 export class DetailPage extends BasePage {
 	constructor(page: Page, domain: string, context: any = null) {
