+++ src/pages/knowledge_app/modules/home_page.ts
@@ -1,0 +1,11 @@
+import { Page, expect } from '@playwright/test';
+
+import { BasePage } from '../../base_page';
+
+export class HomePage extends BasePage {
+	constructor(page: Page, domain: string, context: any = null) {
+		super(page, domain, context);
+	}
+
+	private locators = {};
+}
